Split fuzz-proto to separate package.
