+++ bin/convert/types.ts
@@ -6,7 +6,7 @@ import { blockFromJson, headerFromJson, workReportFromJson } from "@typeberry/bl
 import { type Decode, type Encode, Encoder } from "@typeberry/codec";
 import type { ChainSpec } from "@typeberry/config";
 import { JipChainSpec } from "@typeberry/config-node";
-import { messageCodec } from "@typeberry/ext-ipc/fuzz/v1/types.js";
+import { v1 } from "@typeberry/fuzz-proto";
 import { blake2b } from "@typeberry/hash";
 import type { FromJson } from "@typeberry/json-parser";
 import { decodeStandardProgram } from "@typeberry/pvm-spi-decoder";
@@ -148,8 +148,8 @@ export const SUPPORTED_TYPES: readonly SupportedType[] = [
   },
   {
     name: "fuzz-message",
-    encode: messageCodec,
-    decode: messageCodec,
+    encode: v1.messageCodec,
+    decode: v1.messageCodec,
   },
 ];
