+++ extensions/ipc/package.json
@@ -7,8 +7,9 @@
     "@typeberry/block": "*",
     "@typeberry/bytes": "*",
     "@typeberry/codec": "*",
-    "@typeberry/config": "*",
     "@typeberry/collections": "*",
+    "@typeberry/config": "*",
+    "@typeberry/fuzz-proto": "*",
     "@typeberry/hash": "*",
     "@typeberry/jamnp-s": "*",
     "@typeberry/logger": "*",
