+++ packages/jam/fuzz-proto/index.ts
@@ -1,0 +1,2 @@
+export * from "./server.js";
+export * as v1 from "./v1/index.js";
