+++ packages/jam/node/main-fuzz.ts
@@ -2,7 +2,8 @@ import fs from "node:fs/promises";
 import { type BlockView, Header, type HeaderHash, type StateRootHash, type TimeSlot } from "@typeberry/block";
 import { Bytes } from "@typeberry/bytes";
 import { Encoder } from "@typeberry/codec";
-import { type FuzzVersion, startFuzzTarget, fuzzV1 } from "@typeberry/ext-ipc";
+import { type FuzzVersion, startFuzzTarget } from "@typeberry/ext-ipc";
+import { v1 as fuzzV1 } from "@typeberry/fuzz-proto";
 import { HASH_SIZE } from "@typeberry/hash";
 import { Logger } from "@typeberry/logger";
 import type { StateEntries } from "@typeberry/state-merkleization";
